Write Only supported to perform just allocation before running the OpenCL kernel.
Problem:
   Some applications, such as vector addition, were performing 3 copy_in (input vector
+ output vector). The output vector should be just alloc. This was caused by the Tornado
Data Flow Analyser, in which to determine if it is just write, the analyser checks
for branch divergence. In Tornado, because a previous optmisation, it introduces a new branch
to check the range of the array (and stride with OpenCL).

Solution:
   Check also for the stride. If the IfNode has a check range as a condition, then get this
range and the stride. If the stride is 1, then the param node (output) should be marked as
write only, read_write otherwise.