[FIX] Dynamic switch from a task in a taskscheduler fixed
The problem was mainly in the Tornado Graph. A Tornado Graph looks as follow:

[0]: constant 0
[1]: object 0
[2]: context device=0, [ 3 5 6 ]
[3]: copy in object 0
[4]: dependent write on object 0 by task 5
[5]: task=0, args=[ 3 0 ]
[6]: copy out object 0 after task 5

Where the context is associated with the device. For the first iteration in
the described example, the device is 0. However, for the second iteration,
the context should point to device 1.

[0]: constant 0
[1]: object 0
[2]: context device=1, [ 3 5 6 ]   // Device 1
[3]: copy in object 0
[4]: dependent write on object 0 by task 5
[5]: task=0, args=[ 3 0 ]
[6]: copy out object 0 after task 5

This task should involve recompilation because it is a new device.

protected void scheduleInner() {
        if (result == null) {
            graphContext.assignToDevices();
            compile();
        }
        event = vm.execute();
    }
So the condition for recompilation should also consider if there is a new
device.

This is part of the solution. Then, the Tornado VM assign to each task a stack.
That stack contains the number of arguments and the device context. So every
time there is a new task that needs to be recompiled, the Tornado VM needs
to create a new stack (CallStack) and reset the previous stack.

Github issue #27 and #55