[FEAT] Torando test script documented and stats added.
When passing the unittests, the scripts will print the total number
of thest passed and failed. The tornado-test script is also documented
for future modifications.

This new information should be used to push to the Tornado SDK repository
if all unittests (or at least the subset should always pass) are ok.

Github issue #56