[FIX] GPU Thread selection when having multiple loops
The problem is that when having multiple and not nested loops,
Tornado detects nested parallelism and sets 2D global and local dimensions.
The actual problem is in the TornadoShapeAnalysis.

To fix this, first it explores nested loops. In those are detected, then
it keeps the usual path to eploit 2D and 3D local and global dimensions.
Otherwise, it sets 1D.

Github issue #74