<feat> Multiple reductions in the same task-schedule reusing data supported
When having multiple reduction tasks in the same task-schedule, we need
to stream-in new data only if the input is reused. This is due to the fact
that reductions modifies input/output buffers only on the device side.

To keep consistency, before running consecutive parallel reductions,
we stream-in data only if they are reused.

Github issue #327