[fix] code generator block visitor extra bracket fixed
When traversing the CFG, the code generator was introducing an extra
bracker for some applications.
The error was that the visitor was closing the scope for sections already
closed. This fix traverses all dominators of the current closing block and
checks if they have not been already closed.

Github issue #359