[TP-21] Parallelization of loops with constant loop-bounds
The issue was that for expressions such as "<=", the TornadoTaskSpecialization
phase was inserting the constant node outside the parallel range node.

The solution is that, for the cases in which the Parallel Range contains
a PhiNode as value, we can inspect the condition of the loop-bound. If
a constant node is found, we can perform node swapping between these two.

Github issue #21