[hotfix] Link Narrow Node with the FixedNode for correct Numeric Promotion
The sketcher builds the IR for each reachable method from a given task,
and passes a series of optimizations, starting with Canonicalization.
The first cannicalization phase differs from the Graal execution workflow.
So this commit introduces a new phase to link the new Narrow node with
the store-index. And it works.

In TornadoVM, we need a custom canonicalizer, because we don't want to
simplify vector types, among other types. This path differs from Vanilla-GraalVM
compiler. By doing this, canonicalizer phases in lowering-round start appearing
the sub-canonicalizer phases with the Sign-Extend. THis is what we want, given
correct results.

Github issue #484 & #566