[feat] support parallel hybrid mode execution (CPU-GPU) for reductions
Parallel compilation and execution for hybrid reductions (cpu-gpu).
The main part runs in parallel on the accelerator (e.g., on a GPU) for
input sizes of power of two. The remaning part is specialized and compiled with
Graal and runs on the host single thread.

A final merge is required to merge the results from the accelerator and the
CPU.

Github issue #577