[hotfix] Bug fixed with DEOPT node for Bound Check without array accesses
The problem is related to a Guard Node that is introduced after a
canonicalization phase in the TornadoVM sketcher.
The guard is associated with Bounds Check. However, the array index nodes
are removed.

This patch fixes by traversing the IR to check if the depot is because
there is an active association with an array index load/store or not.
In the case there no association, we can safely remove the node.

Github issue #589