[hotfix] OpenCL code gen fix - basic block traversal
This patch provides a fix for the OpenCL code gen. When genereting an
end-Scope block, we need to check that the current basic block is not
in the true branch and and contains a loop-end. In the case of the
OpenCL code gen, the loop-end is first inspected.

This patch removes the extra close-block for this condition.

Github issue #589