[hotfix] Basic Block Traversal for OpenCL fixed
The issue is that a basic block was scheduled to generate code before
inner loop-if blocks. This block was scheduled as an exit condition using
the true branch. However, to generate correct code in the right-order,
the TornadoVM OCLBlockVisitor must traverse first the inner loops, and
then generate the final exit condition for the first loop.

This patch solves the problem by re-ordering the blocks to be visited
from the OCLCompilationResultBuilder.

Aditionally, we cover some corner cases to not close the brackets from
the merge nodes, if there are pending blocks to be closed.

Github issue #589