[hotfix] Reductions with irregular sizes fixed
The problem was that the hybrid approach of TornadoVM for running reductions
of irregular sizes (not-power of two) use different background Java threads
to compile and execute the remaning code on CPU. The CPU part needs to create
a new thread every time the whole expression is executed.

This is now solve by creating a new execution Java thread for the remaining
part (compilation is cached). Besides, the arrays for the hybrid part need to
be set to the reduce neutral element before running.

How to test it?
```bash
$ tornado uk.ac.manchester.tornado.examples.reductions.ReductionIrregular
```