[fix] Device Information from MxM examples fixed
This fix also provides a patch for quering the device type of the latest
execution.

The problem is that, when setting the device name for a specific task, the
device requested is not the same as the device being queried. This is because
the query happends at the task-schedule level instead of the task-level.

This patch provides fix for querying the device type by returning the device
assigned to the first task. This is currently legal since TornadoVM executes
all task on the same device.

For a multi-task configuration, we will need to pass a task-index when quering
a device assocated to a task within a task-scheduler.

Github issue #539