[feat] TornadoVM support for a single VM per device
This fixes dynamic device switching when having multiple devices of
different types (e.g., CPU, GPU, HD Graphics).
The issue was the the number of thread deployed was not the correct one,
getting only the previous configuration.

This patch extends TornadoVM with a VM instance per device. There is a map
between the contextID from TornadoVM bytecodes and the device selection.

The TornadoTaskScheduler keeps a mapping between VM instances and devices.
[Device01] -> vm01
[Device02] -> vm02
....
[DeviceNN] -> vmNN

Additionally, TornadoVM checks the device using the contextID from the
bytecodes. However the device lists is a one to one mapping between the
phisical information stored in the TornadoBackend and the the VM. This
way it is easier to switch devices based on an index.