[RT] Non-final objects passed via scope not supported
In order to replace the loop bounds from an array-length node to a
constant node and set the parallel dimensions and the number of threads
to be deployed, arrays accessed via scope need to be final.

This patch creates a warning message and a bailout for non-final arrays
passed via scope.