[SPIRV] Unique signature declaration for OpTypeFunctions
To emit a {@link SPIRVOpTypeFunction}, the id has to be unique and the
function return type and the parameters have to be one per function
signature. This means that, if two functions have the same return type and
the same type of parameters and the parameter types are the same. we should
return the same ID previously registered for the first function declaration.

This method checks whether we have a function type with the same return ID
and same parameters already registered. To do so, we manage a table
{@link SPIRVAssembler#opFunctionTable}, that stores the information as
follows:

<code>
    Map<%returnType, Map<NumParameters, LinkedList<SPIRVOpFunctionTable>>>
</code>

If we have the same number of parameters with the same return type, when we
do a sequential search over the linked-list to check the type of each
parameter (stored in the {@link SPIRVOpFunctionTable) class).