[SPIRV] Support for direct calls with vector types.
NOTE: Test to check we can generate vector types for the method signature and
non-main kernel functions.

This is just a test. To store the result, we need to use float4.set(value).
However. this cases an error in lowering. Since these errors are replicated
for all backends, we just check the code gen for vector types, not fot the
access. Future releases will provide store access for non-inlined methods for
all backends.