[SPIRV] OpPhi optimization supported
This commit supports `OpPhi` and Load/Store elimination from the LoopBegin
and LoopEnd Traversal of the GraalIR.

Additionally, to support OpPhi, this commits adds a private table
to be maintained per assembler instance. The main issue with the OpPhi
instruction generation is that, at the time of generating the OpPhi,
some of the Ids (SPIRVIds) might not be known, because the dependent
instruction has not been generated yet.

Additionally, I do not find a way to obtain, from the LIR generation
in Graal, if the current LIR instruction belongs to a phi value or not.
To address these issues, this commit does the following:

A) During LoopBegin in the LIR generation, we know all the PhiValues and
the dependent ValueNodes. Therefore, we introduce a table that relates
an AllocateValue with a new AllocatedValue
           (PhiValue01 -> PhiValueRenamed)
These two values are actually the same value. However, since Graal is SSA,
when we perform the operation that it depends on, we get a new name for it.
Therefore, this table, called phiTrace, will keep a trace of all PhiValues.

B) The LoopEnd will generate the OpPhi, right after the loop label is generated.
The OpPhi ASM instruction will receive a reference to the phiTable and the
phiTrace.

C) During the LIR LoopEnd, we know the final name for the variable of the PhiValue.
Then, at this point, we update the phiTrace.

D) After this process, the ASM code gen starts. Now, we know all phi variables
and the place, and , with the help of the phiTable, we can pass the SPIRVIds
that are generated during the OpPhi. Then, for all operarions (this commit
only provides for SPIRVBinary operations as a Proof of Concept), we check the
`phiTable` and check if the ID has been already generated. In that case, we
obtained it from the table.

An example is as follows:

```asm
%B1_kernel0 = OpLabel
        %54 = OpPhi %uint {%51 %B0_kernel0} {%53 %B1_kernel0}  # ID 53 is generated
        %55 = OpSLessThan %bool %54 %25
              OpBranchConditional %55 %B2_kernel0 %B3_kernel0
%B2_kernel0 = OpLabel
        %58 = OpSConvert %ulong %54
        %60 = OpShiftLeftLogical %ulong %58 %59
        %61 = OpIAdd %ulong %60 %23
        %62 = OpIAdd %ulong %48 %61
        %65 = OpConvertUToPtr %ptr_uint %62
              OpStore %65 %26 Aligned 4
        %66 = OpLoad %v_ulong_3 %spirv_BuiltInGlobalSize Aligned 32
        %67 = OpCompositeExtract %ulong %66 0
        %68 = OpUConvert %uint %67
        %53 = OpIAdd %uint %68 %54          ## ID 53 obtained from the phiTable
              OpBranch %B1_kernel0
%B3_kernel0 = OpLabel
              OpBranch %return_kernel0
```