[SPIRV] Switch and complex CF expressions with opt L/S
This commit supports optimizations of load/stores for switch statements
and complex control flow expressions.

In the case we have an nested-if condition, we can capture the corres-
ponding Phi values through the MergeNode and PhiNode from the Graal LIR.
In the case a merge is present, we annotate the phi values in the
`phiTrace` table and forward the value (pass the value).
When we merge the two values coming from different branches,
we insert an OpPhi instruction with the SPIRV-IDs already generated.

In this case, all SPIRV-IDs were already generated by previous operations
since the new OpPhi will appear in a new basic block in order to perform
the final merge.