[SPIRV] Multiple OpPhi variable generation
There are two main changes from the initial implementation of the OpPhi and
Load/Store elimination instructions:

A) Support for multiple OpPhi instructions in the loop_header block.

In the visitor method of `SPIRVNodeLIRBuilderNode#emitLoopBegin` we need to
detect all phi values that the loop has attached and emit an OpPhi
instruction for all of them. At first it was only passing one phi value.

B) Carry value from a Phi Value of another block.

This is more complicated to see. The case we have is as follows:

```
  phiVal1  = OpPhi …. %fma Loop_Body
LoopBody Label

        …
        %fma = OpExtInst …
        branch loop_header
LoopEnd

        OpStore %fma  ### << THIS IS WRONG. We need to do the store
                            for the %phiVal1, instead of the $fma
```

To support this case, we extended the OpStore code generator in TornadoVM to
check weather the variable to be stored is present in the PhiAcrossBlocks Table
(this is also a new table stored in the Assembler. Whenever there is a new
OpPhi instruction, we update this table for future lookups. In the store
method, we check with this table before storing a value.