[fix] WarmUp using the KernelContext API for SPIR-V
This patch fixes the JIT compilation when invoking the warmUp function
for the SPIR-V backend. The issue in was the detection of a parallel kernel.

In the case of the SPIR-V, we keep a mapping between SPIRVIds and the
intrinsic. If the kernel is not parallel (during the warm-up phase
we do not have the GridScheduler), then these SPIRVIds were not
generated.

This PR simplifies the process of populating the Mapping Table
for the Thread Intrinsics (parallel kernels). We do not need to
check if the kernel is parallel or not. The code generator
checks if there are Thread IDS Nodes, Block-Size, etc. Therefore
the check for the parallel kernel can be safely removed.

All unittests passing.