[SPIRV] Implement Load/Store for Vector Collections
This PR provides a new patch for running TornadoVM collections on
GPUs using the SPIR-V backend.

The VLoad is implemented as follows:

As an example:

```asm
    %301 = OpInBoundsPtrAccessChain %_ptr_Function_double %spirv_d_1F0 %ulong_0 %ulong_0
    %302 = OpPtrCastToGeneric %_ptr_Generic_double %301
    %303 = OpExtInst %v2double %1 vloadn %ulong_0 %302 2
       OpStore %spirv_v2d_50F0 %303 Aligned 16
```

The address to load corresponds to a vector declared in private memory.
Then we need to use cast to generic and call the vload intrinsic with
the right position. The position is computed in the ArithmeticTool
using the vector type and the current position from the IR.

The VSTORE is similarly implemented using the vtore SPIR-V intrinsic
instead.