[SPIRV] Fix the PHI value dependency for binary operations
The problem was that some blocks may depend on a Phi Value declared
in the dominator block. For instance:

```asm
    %B1 = OpLabel
    %77 = OpPhi %uint %uint_0 %B0 %76 %B2
          ...
    %B2 = OpLabel
          ...
    %76 = OpIAdd %uint %77 %uint_1
          ...
    %B3 = OpLabel
    %97 = OpIEqual %bool %76 %73    << Taking the ID 76 instead of 77 (Phi)
          OpBranchConditional %97 %B4 %B5
```

The spirv-val produces the following error:

```bash
ID 76[%76] defined in block 75[%B2] does not dominate its use in block 84[%B3]
  %B3 = OpLabel
```

The solution is to check, when the optimizer is ON, that the result ID for the
binary operations does not correspond to a Phi Value already present in the
forward Table.

After the fix:

```asm
    %B1 = OpLabel
    %77 = OpPhi %uint %uint_0 %B0 %76 %B2
          ...
    %B2 = OpLabel
          ...
    %76 = OpIAdd %uint %77 %uint_1
          ...
    %B3 = OpLabel
    %97 = OpIEqual %bool %77 %73
          OpBranchConditional %97 %B4 %B5
```