[SPIRV][optimizer] Fix for the Phi->Phi variables
This patch provides a fix for the Phi->Phi variables when using the
SPIR-V optimizer. The issue was that for some variables, the Phi
dependency could be given by another Phi variable instead of a
compute instruction.

This patch also takes into account the reusability of SPIR-V Ids
in the case we find a Phi variable in the internal PhiMap.