[SPIRV][optimizer] Phi branch ID dependencies fixed
Since we dynamically assign a new ID that depdends on the forward variables
for Phi instructions, we need to check that the dependent branches do not
point to the same ID. In such cases we need to reassign one of the branches
to the most recent ID.

Besides, since the ID assignation is dynamic, we need to check that
the requested IDs are unique. In the case they are not unique, the JIT
compiler will request a new ID to the SPIRV module.

The sitation in which the IDs might repeat is that Phi values point to
another phi value, and the some of the operands are taken from Phi
instructions. In this case, we request a new ID for the OpPhi Instruction.