[SPIRV] Duplicated SPIRV Ids fixed for non-inlined methods
This patch fixes duplicated IDs for non-inlined methods using the SPIR-V
backend. The problem was the the code generator was annotating multiple
times the same method. This patch fixes the problem by adding a pending
method to a queue only when the method is not found in a Symbol Table.