[SPIRV] Result type for Vector Select Fixed
This patch provides a fix for the result type of vector select operations
in SPIR-V.

The issue was that, when having a VSelect, the generated code was as follows:

```asm
         %89 = OpLoad %v4float %spirv_v4f_10F0 Aligned 16
         %90 = OpCompositeExtract %float %89 1
               OpStore %spirv_v4f_13F0 %90 Aligned 4  << Wrong Type
```

The propagated type for the Store operation was assiged to a vector type,
instead of an scalar operation. This was due to the wrong type setter
for the `SPIRVVectorElementSelect` opcode of the SPIR-V backend.

In the case of the SPIR-V, we need both the scalar type (vector element
type) and the vector type in order to compose all operations involved
in the CompositeExtract instructions of SPIRV.

The resulting code is as follows:

```asm
        %89 = OpLoad %v4float %spirv_v4f_10F0 Aligned 16
        %90 = OpCompositeExtract %float %89 1
              OpStore %spirv_f_13F0 %90 Aligned 4
```

Additionally, this patch also fixes the Unary Operation FNegate for
vector types.