[SPIRV] Missing OpPhi instruction for constant values fixed
This patch provides a fix for missing OpPhi instructions when the operands
are of type constant. The problem was the the JIT compiler was checking
for the values to be already registered on the PhiTrace table. However,
for constant values, we never register the entry on that table.

This patch covers this case in order to generate the right OpPhi
instruction with all the operands.