[WIP] OpenCL backend migrated to Graal 22.1.0 compiler
From Graal 22.1.0 the graph traversal was changed. This patch provides
a rescheduler of basic blocks (if-else). Since Graal 22.1.0, it generates
first the false condition rather than the true condition (using ordered IDs).

In an if-else block, the basic block to be generated must be always the true
condition before the false condition only if we have a LoopEndNode node
in the false branch, or we have a LoopExit in the true branch contains
a LoopExitNode or it is not a control Split (due to nested control-flow).