[fix] PTX Thread Block scheduler for 1D reductions
When running TornadoVM using the PTX backend on 30XX GPUs, the thread
block was not calculated properly. For 1D we what to maximize the occuppancy
but the runtime was obtining a lower number of threads compared to the
JIT compiler during the partial evaluation.

This patch provides a fix for this by using the maxThreadBlock instead
of the max occuppancy in the case of 1D kernels.

Related issue #194