[compiler] WRITE-Only values improved in the Data Dependency phase
The TornadoVM JIT compiler is very convervative with the data dependencies
for array accesses. But some codes can be further optimized by just
allocating data and not copying data from the host to the device.
This commit adds an extra check to determine if a store indexed node (array)
is written in all positions of the iteration space. In that case, TornadoVM
sets a WRITE-ONLY for the array of interest.

This optimization saves data transfers from the host to the device.

Related issue: #150