[spirv][fix] Multi-device SPIR-V exploration fixed
This PR solves the issue of `stack smashing detected` during a jni
return call for the LevelZero device exploration.

As stated in the Ubuntu documentation:

> NOTE: In Ubuntu 14.10 and later versions, -fstack-protector-strong is
  enabled by default for C, C++, ObjC, ObjC++, if none of -fno-stack-protector,
  -nostdlib, nor -ffreestanding are found.

As suggested here: https://stackoverflow.com/questions/10712972/what-is-the-use-of-fno-stack-protector

> In the standard/stock GCC, stack protector is off by default.
However, some Linux distributions have patched GCC to turn it on by default.

We need this flag to turn off for the integration with the Level Zero drivers.