[runtime] Profiler fixed for reductions
The issue is that the profiler was NULL due to task rewriting at runtime.
The fix is that we pass the profiler instance along to be setup by
the new execution plan. Besides, the query for all profiling metrics
need to check if there is an expression that was rewritten. In that case,
we return the metric from the new execution plan, rather than the first
execution plan.