[fix] Runtime Name Check for Grid Schedulers
Grid Schedulers contain the name of an existing task graph and a task name.
However, it could be possible that, due to a programmers error, the
grid name does not match with any existing tasks within the
assigned task-graph. This PR adds this check and it refactors
the PrebuilTask API to use the same style as the Compilable Task API.