[feat] Dynanically enable the profiler after the first execution
It should be possible to enable the profiler afte the TornadoVM
Execution Plans runs the first iteration. There was an issue with the
profiler that was preventing the profiler to be registered if it was
not enabled in the first iteration. This commit fixes and enables the
profiler at any point during runtime via the execution plan.

Example:

```java
TornadoExecutionPlan executionPlan = new TornadoExecutionPlan(taskGraph.snapshot());
// Runs without the profiler
executionPlan.execute();

// enable the profiler
TornadoExecutionResult executionResult = executionPlan.withProfiler(ProfilerMode.SILENT)
             .execute();

// Query timers from the profiler
executionResult.getProfilerResult().getDeviceKernelTime();
```