[feat] Java Multi-thread execution plans supported
This commit also improves the TornadoVM Memory Model.
All variable states (method parameters and field acceses) are handled
and managed at the execution plan-level. Thus, it is possible to reuse
the same task-graph for different execution plans and have private
local variable states.
States are defined by STREAM-IN, OUT, and FORCE-IN as well as a
device attached to it.