Fixed CL_DEVICE_TYPE_ALL to be a long type to avoid sign extension
The CL_DEVICE_TYPE_ALL enumeration was coming out as 0xffffffffffffffff
rather than 0xffffffff when used with clGetDeviceIDs which was causing
some OpenCL backends to fail as it was passing a technically incorrect
value.

This resolves this by changing the immediate for CL_DEVICE_TYPE_ALL to
be have the suffix 'L' which avoids the sign extension and removes the
previous workaround for Codeplay as a backend.