[fix] Fix the plugins for the MemorySegment class
The issue was that MemorySegment is an interface, and Graal does not
apply the "InvocationPlugin" unless it has concrete type information
about the receiver object at the call site. "NodePlugins" are always
applied though, so the fix was to switch from using an
"InvocationPlugin" for the MemorySegment interface to "NodePlugins".