[fix] kernel name for the PTX with sanitizer check
Fix issue #663. The problem is that, for the generated PTX kernel,
the task-id is used. If the task-id contains illegal C characters
for a function name, then it does not compile the PTX kernel.

This patch adds a sanitizer to check for non-legal characters and
replace them with another character.